Move bbox_to_native_bbox from layman to geoserver
